Rename `modes` -> `sync` and remove uppercase module names
